현재 프론트엔드 프로젝트 구조 개선 : feat : 불필요한 파일 정리 https://github.com/Sejong-Balsamic/Malsami-FE/issues/433
