+++ src/components/common/tags/YeopjeonTag.tsx
@@ -1,10 +1,10 @@
 import React from "react";
-import ImageWrapper from "@/components/deprecated/ImageWrapper";
+import Image from "next/image";
 
 function YeopjeonTag({ point }: { point: number }) {
   return (
     <span className="font-pretendard-semibold inline-block text-[12px] text-black">
-      <ImageWrapper src="/icons/Yeopjeon.svg" />
+      <Image src="/icons/Yeopjeon.svg" width={14} height={14} alt="Yeopjeon" />
       <span className="ml-1">{point}</span>
     </span>
   );
